ai_call_agent/requirements.txt
fastapi>=0.104.1
uvicorn>=0.24.0
python-dotenv>=1.0.0
python-multipart>=0.0.6
requests>=2.31.0
SpeechRecognition>=3.8.1
PyAudio>=0.2.11
numpy>=1.24.0
websockets>=10.0
starlette>=0.27.0
google-cloud-dialogflow-cx
google-api-core
google-auth
google-cloud-core
google-cloud-speech>=2.21.0 